feat(jsonnet): support setting extVars
`extVars` are a concept of Jsonnet that allow to inject data into the compiler
from the command line.

At Grafana, we have a build-image whose tag must be consistent across all CI
systems, so we use such an extVar for this.

To be able to adopt drone-cli (were using plain `jsonnet` before) we need this
functionality exposed :D